requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0 ; python_version >= '3.8'
certifi==2024.2.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
htmltestrunner-rv==1.1.2
idna==3.6 ; python_version >= '3.5'
jinja2==3.1.3 ; python_version >= '3.7'
markupsafe==2.1.5 ; python_version >= '3.7'
pydantic==2.6.4
pydantic-core==2.16.3 ; python_version >= '3.8'
pyparsing==3.1.2 ; python_full_version >= '3.6.8'
requests==2.31.0
typing-extensions==4.10.0 ; python_version >= '3.8'
urllib3==2.2.1 ; python_version >= '3.8'
